refactor candidate skills route

+++ skill/routes.py
@@ -114,41 +114,41 @@ def get_skills():
 @skill.route("/api/candidate/skills", methods=['GET', 'PUT'])
 @jwt_required()
 def candidate_skills():
-    with connect_db() as connection:
-        cursor = connection.cursor()
+    try:
+        with connect_db() as connection:
+            cursor = connection.cursor()
 
-        # Получение данных о пользователе
-        user_id = get_jwt_identity()
-        cursor.execute("SELECT * FROM users WHERE id = ?", (user_id,))
-        existing_user = cursor.fetchone()
-        columns = [column[0] for column in cursor.description]
-        user = {columns[i]: existing_user[i] for i in range(len(columns))}
-
-        if user['role'] not in [1, 2]:
-            return jsonify({'error': 'Навыки могут менять только соискатели'}), 400
-
-        if request.method == 'GET':
-            cursor.execute("SELECT * FROM user_skill WHERE user_id == ?", (user['id'],))
-
+            # Получение данных о пользователе
+            user_id = get_jwt_identity()
+            cursor.execute("SELECT * FROM users WHERE id = ?", (user_id,))
+            existing_user = cursor.fetchone()
             columns = [column[0] for column in cursor.description]
-            skills = [dict(zip(columns, row)) for row in cursor.fetchall()]
-            return jsonify({'skills': skills})
-        else:
-            data = request.get_json()
+            user = {columns[i]: existing_user[i] for i in range(len(columns))}
 
-            # Проверка на наличие массива навыков теле запроса
-            error_response = required_fields(data, ['skills'])
-            if error_response:
-                return jsonify({'error': 'В теле нет массива навыков'}), 400
+            if user['role'] not in [1, 2]:
+                return jsonify({'error': 'Навыки могут менять только соискатели'}), 403
 
-            cursor.execute('''
-                        DELETE FROM user_skill
-                        WHERE user_id = ? AND skill_id NOT IN ({})
-                    '''.format(','.join(['?'] * len(data['skills']))), [user['id']] + data['skills'])
+            if request.method == 'GET':
+                cursor.execute("SELECT * FROM user_skill WHERE user_id == ?", (user['id'],))
 
-            for skill_id in data['skills']:
-                cursor.execute('''
-                            INSERT INTO user_skill (user_id, skill_id)
-                            VALUES (?, ?)
-                            ON CONFLICT(user_id, skill_id) DO NOTHING
-                        ''', (user['id'], skill_id))
+                columns = [column[0] for column in cursor.description]
+                skills = [dict(zip(columns, row)) for row in cursor.fetchall()]
+                return jsonify({'skills': skills})
+            else:
+                data = request.get_json()
+
+                # Проверка на наличие массива навыков теле запроса
+                error_response = required_fields(data, ['skills'])
+                if error_response:
+                    return jsonify({'error': 'В теле нет массива навыков'}), 400
+
+                # Удаляем записи про старые навыки
+                cursor.execute("DELETE FROM user_skill WHERE user_id = ?", (user['id']))
+
+                # Добавляем новые
+                for skill_id in data['skills']:
+                    cursor.execute("INSERT INTO user_skill (user_id, skill_id) VALUES (?, ?)", (user['id'], skill_id))
+
+    except Exception as e:
+        print(f"Ошибка подключения к базе данных: {e}")
+        return jsonify({'error': 'Ошибка сервера'}), 500
